requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# HTTP and networking
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pyyaml>=6.0.1

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1.0
cryptography>=41.0.0

# QR code generation
qrcode[pil]>=7.4.2

# Background tasks
apscheduler>=3.10.4

# Production server
gunicorn>=21.2.0
